+++ src/Stack/Package.hs
@@ -51,7 +51,6 @@ import qualified Data.Set as S
 import qualified Data.Text as T
 import           Data.Text.Encoding (decodeUtf8, decodeUtf8With)
 import           Data.Text.Encoding.Error (lenientDecode)
-import           Data.Version (showVersion)
 import           Distribution.Compiler
 import           Distribution.ModuleName (ModuleName)
 import qualified Distribution.ModuleName as Cabal
@@ -65,7 +64,13 @@ import           Distribution.ParseUtils
 import           Distribution.Simple.Utils
 import           Distribution.System (OS (..), Arch, Platform (..))
 import qualified Distribution.Text as D
+import qualified Distribution.Types.CondTree as Cabal
+import qualified Distribution.Types.Dependency as Cabal
+import qualified Distribution.Types.ExeDependency as Cabal
+import qualified Distribution.Types.LegacyExeDependency as Cabal
+import qualified Distribution.Types.UnqualComponentName as Cabal
 import qualified Distribution.Verbosity as D
+import           Distribution.Version (showVersion)
 import qualified Hpack
 import qualified Hpack.Config as Hpack
 import           Path as FL
@@ -116,7 +121,7 @@ readPackageUnresolvedBS source bs =
 rawParseGPD :: BS.ByteString
             -> Either PError ([PWarning], GenericPackageDescription)
 rawParseGPD bs =
-    case parsePackageDescription chars of
+    case parseGenericPackageDescription chars of
        ParseFailed per -> Left per
        ParseOk warnings gpkg -> Right (warnings,gpkg)
   where
@@ -217,7 +222,9 @@ packageFromPackageDescription packageConfig pkgFlags pkg =
     , packageLicense = license pkg
     , packageDeps = deps
     , packageFiles = pkgFiles
-    , packageTools = packageDescTools pkg
+    , packageTools = map
+        (\(Cabal.ExeDependency name' _ range) -> Cabal.Dependency name' range)
+        (packageDescTools pkg)
     , packageGhcOptions = packageConfigGhcOptions packageConfig
     , packageFlags = packageConfigFlags packageConfig
     , packageDefaultFlags = M.fromList
@@ -225,13 +232,14 @@ packageFromPackageDescription packageConfig pkgFlags pkg =
     , packageAllDeps = S.fromList (M.keys deps)
     , packageHasLibrary = maybe False (buildable . libBuildInfo) (library pkg)
     , packageTests = M.fromList
-      [(T.pack (testName t), testInterface t) | t <- testSuites pkg
+      [(T.pack (Cabal.unUnqualComponentName $ testName t), testInterface t) | t <- testSuites pkg
                                               , buildable (testBuildInfo t)]
     , packageBenchmarks = S.fromList
-      [T.pack (benchmarkName biBuildInfo) | biBuildInfo <- benchmarks pkg
+      [T.pack (Cabal.unUnqualComponentName $ benchmarkName biBuildInfo) | biBuildInfo <- benchmarks pkg
                                           , buildable (benchmarkBuildInfo biBuildInfo)]
     , packageExes = S.fromList
-      [T.pack (exeName biBuildInfo) | biBuildInfo <- executables pkg
+      [T.pack (Cabal.unUnqualComponentName $ exeName biBuildInfo)
+        | biBuildInfo <- executables pkg
                                     , buildable (buildInfo biBuildInfo)]
     -- This is an action used to collect info needed for "stack ghci".
     -- This info isn't usually needed, so computation of it is deferred.
@@ -338,19 +346,19 @@ generatePkgDescOpts sourceMap installedMap omitPkgs addPkgs cabalfp pkg componen
                    , fmap
                          (\exe ->
                                generate
-                                    (CExe (T.pack (exeName exe)))
+                                    (CExe (T.pack (Cabal.unUnqualComponentName (exeName exe))))
                                     (buildInfo exe))
                          (executables pkg)
                    , fmap
                          (\bench ->
                                generate
-                                    (CBench (T.pack (benchmarkName bench)))
+                                    (CBench (T.pack (Cabal.unUnqualComponentName (benchmarkName bench))))
                                     (benchmarkBuildInfo bench))
                          (benchmarks pkg)
                    , fmap
                          (\test ->
                                generate
-                                    (CTest (T.pack (testName test)))
+                                    (CTest (T.pack (Cabal.unUnqualComponentName (testName test))))
                                     (testBuildInfo test))
                          (testSuites pkg)]))
   where
@@ -531,13 +539,13 @@ packageDependencies pkg =
 packageToolDependencies :: PackageDescription -> Map Text VersionRange
 packageToolDependencies =
   M.fromList .
-  concatMap (fmap (packageNameText . depName &&& depRange) .
+  concatMap (fmap (\(Cabal.LegacyExeDependency name range) -> (T.pack name, range)) .
              buildTools) .
   allBuildInfo'
 
 -- | Get all dependencies of the package (buildable targets only).
-packageDescTools :: PackageDescription -> [Dependency]
-packageDescTools = concatMap buildTools . allBuildInfo'
+packageDescTools :: PackageDescription -> [Cabal.ExeDependency]
+packageDescTools = concatMap buildToolDepends . allBuildInfo'
 
 -- | This is a copy-paste from Cabal's @allBuildInfo@ function, but with the
 -- @buildable@ test removed. The implementation is broken.
@@ -551,12 +559,10 @@ allBuildInfo' pkg_descr = [ bi | Just lib <- [library pkg_descr]
                               , True || buildable bi ]
                       ++ [ bi | tst <- testSuites pkg_descr
                               , let bi = testBuildInfo tst
-                              , True || buildable bi
-                              , testEnabled tst ]
+                              , True || buildable bi ]
                       ++ [ bi | tst <- benchmarks pkg_descr
                               , let bi = benchmarkBuildInfo tst
-                              , True || buildable bi
-                              , benchmarkEnabled tst ]
+                              , True || buildable bi ]
 
 -- | Get all files referenced by the package.
 packageDescModulesAndFiles
@@ -596,9 +602,9 @@ packageDescModulesAndFiles pkg = do
     return (modules, files, dfiles, warnings)
   where
     libComponent = const CLib
-    exeComponent = CExe . T.pack . exeName
-    testComponent = CTest . T.pack . testName
-    benchComponent = CBench . T.pack . benchmarkName
+    exeComponent = CExe . T.pack . Cabal.unUnqualComponentName . exeName
+    testComponent = CTest . T.pack . Cabal.unUnqualComponentName . testName
+    benchComponent = CBench . T.pack . Cabal.unUnqualComponentName . benchmarkName
     asModuleAndFileMap label f lib = do
         (a,b,c) <- f lib
         return (M.singleton (label lib) a, M.singleton (label lib) b, c)
@@ -651,7 +657,7 @@ resolveGlobFiles =
 --
 matchDirFileGlob_ :: (MonadLogger m, MonadIO m) => String -> String -> m [String]
 matchDirFileGlob_ dir filepath = case parseFileGlob filepath of
-  Nothing -> liftIO $ die $
+  Nothing -> liftIO $ throwString $
       "invalid file glob '" ++ filepath
       ++ "'. Wildcards '*' are only allowed in place of the file"
       ++ " name, not in the directory name or file extension."
@@ -681,7 +687,7 @@ benchmarkFiles bench = do
     dir <- asks (parent . fst)
     (modules,files,warnings) <-
         resolveFilesAndDeps
-            (Just $ benchmarkName bench)
+            (Just $ Cabal.unUnqualComponentName $ benchmarkName bench)
             (dirs ++ [dir])
             (bnames <> exposed)
             haskellModuleExts
@@ -705,7 +711,7 @@ testFiles test = do
     dir <- asks (parent . fst)
     (modules,files,warnings) <-
         resolveFilesAndDeps
-            (Just $ testName test)
+            (Just $ Cabal.unUnqualComponentName $ testName test)
             (dirs ++ [dir])
             (bnames <> exposed)
             haskellModuleExts
@@ -730,7 +736,7 @@ executableFiles exe = do
     dir <- asks (parent . fst)
     (modules,files,warnings) <-
         resolveFilesAndDeps
-            (Just $ exeName exe)
+            (Just $ Cabal.unUnqualComponentName $ exeName exe)
             (dirs ++ [dir])
             (map DotCabalModule (otherModules build) ++
              [DotCabalMain (modulePath exe)])
@@ -782,7 +788,7 @@ targetJsSources = jsSources
 resolvePackageDescription :: PackageConfig
                           -> GenericPackageDescription
                           -> PackageDescription
-resolvePackageDescription packageConfig (GenericPackageDescription desc defaultFlags mlib exes tests benches) =
+resolvePackageDescription packageConfig (GenericPackageDescription desc defaultFlags mlib _subLibs _foreignLibs exes tests benches) =
   desc {library =
           fmap (resolveConditions rc updateLibDeps) mlib
        ,executables =
@@ -811,12 +817,18 @@ resolvePackageDescription packageConfig (GenericPackageDescription desc defaultF
                  (buildInfo exe) {targetBuildDepends = deps}}
         updateTestDeps test deps =
           test {testBuildInfo =
-                  (testBuildInfo test) {targetBuildDepends = deps}
-               ,testEnabled = packageConfigEnableTests packageConfig}
+                  (testBuildInfo test)
+                    { targetBuildDepends = deps
+                    , buildable = packageConfigEnableTests packageConfig
+                    }
+               }
         updateBenchmarkDeps benchmark deps =
           benchmark {benchmarkBuildInfo =
-                       (benchmarkBuildInfo benchmark) {targetBuildDepends = deps}
-                    ,benchmarkEnabled = packageConfigEnableBenchmarks packageConfig}
+                       (benchmarkBuildInfo benchmark)
+                         { targetBuildDepends = deps
+                         , buildable = packageConfigEnableBenchmarks packageConfig
+                         }
+                    }
 
 -- | Make a map from a list of flag specifications.
 --
@@ -854,7 +866,7 @@ resolveConditions :: (Monoid target,Show target)
 resolveConditions rc addDeps (CondNode lib deps cs) = basic <> children
   where basic = addDeps lib deps
         children = mconcat (map apply cs)
-          where apply (cond,node,mcs) =
+          where apply (Cabal.CondBranch cond node mcs) =
                   if condSatisfied cond
                      then resolveConditions rc addDeps node
                      else maybe mempty (resolveConditions rc addDeps) mcs
@@ -1256,10 +1268,10 @@ cabalFilePackageId
     :: (MonadIO m, MonadThrow m)
     => Path Abs File -> m PackageIdentifier
 cabalFilePackageId fp = do
-    pkgDescr <- liftIO (D.readPackageDescription D.silent $ toFilePath fp)
+    pkgDescr <- liftIO (D.readGenericPackageDescription D.silent $ toFilePath fp)
     (toStackPI . D.package . D.packageDescription) pkgDescr
   where
-    toStackPI (D.PackageIdentifier (D.PackageName name) ver) = do
+    toStackPI (D.PackageIdentifier (D.unPackageName -> name) ver) = do
         name' <- parsePackageNameFromString name
         ver' <- parseVersionFromString (showVersion ver)
         return (PackageIdentifier name' ver')